requirements.txt
 
websocket-client
pandas
numpy
scikit-learn
streamlit
matplotlib
asyncio
